Moved the following functions into EWSApiModule:
- get_folder
    - folder_to_context_entry
- get_expanded_group
    - ExpandGroup class
- mark_item_as_read
